feat[#2]: Finora 대시보드 컴포넌트 구조화 및 UI/UX 개선
- 대시보드를 Overview/Transactions/Alerts 탭으로 모듈화
- Aurora 테마 기반 현대적 디자인 적용
- URL 동기화, 다크모드, 실시간 알림 기능 구현
- 재사용 가능한 UI 컴포넌트 라이브러리 구축
- TypeScript 타입 정의 및 유틸리티 함수 분리
